week5-voice-chatbot-implementation/requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0
requests>=2.31.0
python-dotenv>=1.0.0
